Adds roles for games; Removes rolepicker channel id restriction
In the previous commits it was not possible to send the rolepicker messages in multiple channels as each new message would assign the new channel to be the main rolepicker message. The reaction listenner would not care, but the verification approval message asking users to add roles would be inconsistent. However, it is still the intended behavior that certain roles should be listed under one channeel, so message-include.php will still does thie with the exception of messages intended for submenus such as those that would require a previous role to add (IE. Deep Rock Galactic would be posted in the games rolepicker which requires having the Games role first, but the Games role can be removed later and the Deep Rock Galactic role would still function, although the category and games rolepicker channel would no longer be visible)