x2
The original purpose of this function was to prevent the processing of a chat message where the author part was a "user" instead of a "member" as this was either caused by an emptied members repository or a direct message. The issue that caused the members repository to be emptied is resolved and the author part should always be a member so long as the message was sent in a guild AND the member intent is enabled on the discord developer portal.