Added Streamers and Viewers role to Game Roles; Member Update Fix; New Custom Roles default template
Palace Bot does not use the Twitch API to determine if someone goes live and currently does not capture any related info from Discord, so if you want that type of feature you should add another bot such as YAGPDB.xyz, which could use the Streamers role to determine whose streams to announce and the message could be broadcasted to a channel only Viewers have access to.

GUILD_MEMBER_UPDATE has been altered to not spam the logs every time it 'sees' a new member for the first time if they already exist in a guild but loadAllMembers wasn't enabled when the bot was launched, thus thinking a new nickname or a bunch of roles were being added to someone.

Added a more relevant Custom Roles default. The old one was meant for the Civ13 server whom requested it originally.