Message reaction refactoring; Rolepicker rework
New functionality makes it so that adding a reaction adds a role and removing a reaction removes a role. The same rules apply as before, the corresponding role option must be enabled before the bot will attempt to either add, remove, or create roles, and the best practice is to enable the role before creating the role reaction message so that all relevant roles are created at the same time.